requirements.txt
numpy==1.19.2
opencv-python==4.5.5.62
tqdm==4.54.0
imutils==0.5.3
Pillow==9.0.1
